Models at different places, now work together
The problem now is that when we have a request waiting for data
and something else relies on that data, when the data is loaded,
the waiting party doesn't get a full working model. It gets the
data but no binding.

Sort of:

var1 = dataService.get({'roaster':'1'});
var2 = dataStore.get('roaster','1');

var2 will get the data loaded but the model has no bindings.