Updated to work with angularjs 1.2_rc3
There is a new way promises are handled with regards to how data
is returned from a fulfilled promise. Or rather it isn't.

We are using the workaround with $promiseProvider.unwrapPromises
for the time being but this is something that has to change.

I guess we should move away from

var x  = dataStore.get(....)

to

$scope.something = dataStore.get(....)
$scope.something.then(function(data)
 $scope.something = data;
);

or something